[ change ] - move tests into correct tree folder
