+++ test/src/scripts/tableService.test.js
@@ -1,9 +1,9 @@
 'use strict';
 
-jest.mock('./src/scripts/tableService');
+jest.mock('../../../src/scripts/tableService');
 
-const TableService = require('./src/scripts/tableService');
-const Character = require('./src/scripts/character');
+const TableService = require('../../../src/scripts/tableService');
+const Character = require('../../../src/scripts/character');
 const CLASS_NAME = 'table';
 
 
@@ -11,3 +11,4 @@ test('Should prepare environment without errors', () => {
   document.body.innerHTML = `<div class="${CLASS_NAME}"></div>`;
   new TableService(CLASS_NAME);
 });
+
